# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AqariTalk API — AI-guided real estate workflow platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact-release:
    post:
      operationId: requestContactRelease
      tags: [contact-release]
      summary: Buyer requests contact release for a property
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactReleaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRelease"
  /contact-release/{id}:
    get:
      operationId: getContactRelease
      tags: [contact-release]
      summary: Get a contact release record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact release record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRelease"
  /contact-release/{id}/acknowledge:
    post:
      operationId: acknowledgeContactRelease
      tags: [contact-release]
      summary: Buyer or seller acknowledges commission terms
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckInput"
      responses:
        "200":
          description: Updated contact release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRelease"
  /contact-release/property/{propertyId}:
    get:
      operationId: getContactReleaseByProperty
      tags: [contact-release]
      summary: Get contact release for a specific property (for current user)
      security:
        - bearerAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact release or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactReleaseOrNull"

  # ── Commission ────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Marked read

  # ── Admin ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactRelease:
      type: object
      required: [id, buyerId, sellerId, propertyId, status, commissionBuyerPct, commissionSellerPct, createdAt]
      properties:
        id:
          type: string
        buyerId:
          type: string
        sellerId:
          type: string
        propertyId:
          type: string
        status:
          type: string
          enum: [pending, buyer_acked, seller_acked, released, expired]
        commissionBuyerPct:
          type: number
        commissionSellerPct:
          type: number
        buyerAckAt:
          type: ["string", "null"]
          format: date-time
        sellerAckAt:
          type: ["string", "null"]
          format: date-time
        releasedAt:
          type: ["string", "null"]
          format: date-time
        buyerPhone:
          type: ["string", "null"]
          description: Only present when status = released
        sellerPhone:
          type: ["string", "null"]
          description: Only present when status = released
        createdAt:
          type: string
          format: date-time
    ContactReleaseOrNull:
      oneOf:
        - $ref: "#/components/schemas/ContactRelease"
        - type: "null"
    ContactReleaseInput:
      type: object
      required: [propertyId]
      properties:
        propertyId:
          type: string
    AckInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [buyer, seller]

    # ── Commission schemas ────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, titleAr, titleEn, bodyAr, bodyEn, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        bodyAr:
          type: string
        bodyEn:
          type: string
        readAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ── Admin schemas ─────────────────────────────────
...